Release/1.7.1

+++ application/src/main/java/com/mercadolibre/planning/model/me/gateways/staffing/dtos/response/Operation.java
@@ -7,5 +7,5 @@ public class Operation {
 
     final String alias;
 
-    final Double result;
+    final Integer result;
 }
